requirements.txt
matplotlib
psutil
tqdm
pandas
scipy
numpy
ipykernel
azureml-sdk
azureml-core
azureml-defaults
azureml-mlflow
azureml-telemetry
scikit-learn
debugpy
pyodbc
openpyxl
azure-ai-ml
mltable